refactor(openomni): honest verdict record, shared resolver (#606)
